UX design

+++ post-it/components/ProjectList.js
@@ -1,14 +1,13 @@
+// npm install react-native-super-grid
 import React, { useEffect, useState } from "react";
-import {
-  View,
-  Text,
-  TouchableOpacity,
-  StyleSheet,
-  Image,
-  FlatList,
-} from "react-native";
-import { getProjects, deleteProject, getUserData } from "../API/postItAPI";
+import { View, Text, TouchableOpacity, StyleSheet, Alert } from "react-native";
+import { getProjects, deleteProject } from "../API/postItAPI";
 import { useNavigation } from "@react-navigation/native";
+import { FlatGrid } from "react-native-super-grid";
+import { FontAwesomeIcon } from "@fortawesome/react-native-fontawesome";
+import { faPlus } from "@fortawesome/free-solid-svg-icons/faPlus";
+import { faArrowsRotate } from "@fortawesome/free-solid-svg-icons/faArrowsRotate";
+import { faTrash } from "@fortawesome/free-solid-svg-icons/faTrash";
 
 /**
  * Composant permettant de récupérrer et d'afficher les projets auquels un utilisateur est assigné.
@@ -38,40 +37,19 @@ export default function ProjectList({ username, token, userRole }) {
     callback(username, token);
   }, [username, token]);
 
-  /**
-   * Fonction permettant de récuperer l'id d'un utilisateur grâce à getUserData, provenant de ../API/postitApi.js
-   * et de mettre cet id dans une variable 'userId'.
-   * @param {String} username Le nom d'utilisateur connecté
-   * @param {String} token    Le token d'autorisation de l'api
-   */
-  const getId = (username, token) => {
-    getUserData(username, token).then((data) => {
-      setUserId(data[0].id);
-      console.log(data[0].id);
-    });
-  };
-
-  useEffect(() => {
-    getId(username, token);
-  }, [username, token]);
-
-  const clickHandler = () => {
-    //function to handle click on floating Action Button
-    alert("Todo : page de création de projet !");
-  };
-
   if (userRole == "manager") {
     return (
       <>
-        <View>
-          <Text style={styles.text}>ID de l'user : {userId}</Text>
-          {console.log(projects)}
+        <View style={styles.container}>
           <Text style={styles.text}>Liste des Projets :</Text>
-          <FlatList
-            style={{ textAlign: "left", paddingLeft: 10, paddingTop: 20 }}
+          <FlatGrid
             data={projects}
+            style={styles.gridView}
+            spacing={10}
             renderItem={({ item }) => (
-              <View style={{ flexDirection: "row" }}>
+              <View
+                style={[styles.itemContainer, { backgroundColor: "#51557E" }]}
+              >
                 <TouchableOpacity
                   onPress={() => {
                     // On navigue vers la liste de posts en passant l'id et le titre du projet en paramètres
@@ -89,35 +67,54 @@ export default function ProjectList({ username, token, userRole }) {
                     {item.title}
                   </Text>
                 </TouchableOpacity>
+                {/* 
+                La suppression fonctionne avec un alert qui ne fonctionne que sur téléphone
+                Il est impossible d'utiliser cette fonction sur Web !
+                */}
                 <TouchableOpacity
                   onPress={async (e) => {
                     e.preventDefault();
-                    console.log("suppression de " + item.title + " " + item.id);
-                    await deleteProject(item.id, item.title, token);
-                    callback(username, token);
+                    Alert.alert(
+                      "Suppression",
+                      "Voulez vous supprimer " + item.title + " ?",
+                      [
+                        {
+                          text: "Annuler",
+                        },
+                        {
+                          text: "Oui",
+                          onPress: async () => {
+                            await deleteProject(item.id, item.title, token);
+                            callback(username, token);
+                          },
+                        },
+                      ]
+                    );
                   }}
                 >
-                  <Image
-                    source={require("../assets/trash-can-outline-white.png")}
-                    style={{ height: 24, width: 24 }}
-                  />
+                  <FontAwesomeIcon icon={faTrash} color="#D6D5A8" size={15} />
                 </TouchableOpacity>
               </View>
             )}
           />
         </View>
-
+        <TouchableOpacity
+          activeOpacity={0.7}
+          onPress={() => {
+            callback(username, token);
+          }}
+          style={styles.refreshTouchableOpacityStyle}
+        >
+          <FontAwesomeIcon icon={faArrowsRotate} color="#D6D5A8" size={35} />
+        </TouchableOpacity>
         <TouchableOpacity
           activeOpacity={0.7}
           onPress={() => {
             navigation.navigate("Nouveau Projet");
           }}
-          style={styles.touchableOpacityStyle}
+          style={styles.plusTouchableOpacityStyle}
         >
-          <Image
-            source={require("../assets/plus_icon_repaint.png")}
-            style={styles.floatingButtonStyle}
-          />
+          <FontAwesomeIcon icon={faPlus} color="#D6D5A8" size={35} />
         </TouchableOpacity>
       </>
     );
@@ -125,36 +122,73 @@ export default function ProjectList({ username, token, userRole }) {
 
   if (userRole == "writter") {
     return (
-      <View>
-        <Text style={styles.text}>ID de l'user : {userId}</Text>
-        {console.log(projects)}
-        <Text style={styles.text}>Liste des Projets :</Text>
-        <FlatList
-          style={{ textAlign: "center", paddingLeft: 10, paddingTop: 20 }}
-          data={projects}
-          renderItem={({ item }) => (
-            <View>
-              <TouchableOpacity
-                onPress={() => {
-                  // On navigue vers la liste de posts en passant l'id et le titre du projet en paramètres
-                  navigation.navigate("Posts", {
-                    id: item.id,
-                  });
-                }}
+      <>
+        <View style={styles.container}>
+          <Text style={styles.text}>Liste des Projets :</Text>
+          <FlatGrid
+            data={projects}
+            style={styles.gridView}
+            spacing={10}
+            renderItem={({ item }) => (
+              <View
+                style={[styles.itemContainer, { backgroundColor: "#51557E" }]}
               >
-                <Text
-                  style={[
-                    styles.text_item,
-                    { color: "#D6D5A8", textDecorationLine: "underline" },
-                  ]}
+                <TouchableOpacity
+                  onPress={() => {
+                    // On navigue vers la liste de posts en passant l'id et le titre du projet en paramètres
+                    navigation.navigate("Posts", {
+                      id: item.id,
+                    });
+                  }}
                 >
-                  {item.title}
-                </Text>
-              </TouchableOpacity>
-            </View>
-          )}
-        />
-      </View>
+                  <Text
+                    style={[
+                      styles.text_item,
+                      { color: "#D6D5A8", textDecorationLine: "underline" },
+                    ]}
+                  >
+                    {item.title}
+                  </Text>
+                </TouchableOpacity>
+              </View>
+            )}
+          />
+          {/* <FlatList
+            style={{ textAlign: "center", paddingLeft: 10, paddingTop: 20 }}
+            data={projects}
+            renderItem={({ item }) => (
+              <View>
+                <TouchableOpacity
+                  onPress={() => {
+                    // On navigue vers la liste de posts en passant l'id et le titre du projet en paramètres
+                    navigation.navigate("Posts", {
+                      id: item.id,
+                    });
+                  }}
+                >
+                  <Text
+                    style={[
+                      styles.text_item,
+                      { color: "#D6D5A8", textDecorationLine: "underline" },
+                    ]}
+                  >
+                    {item.title}
+                  </Text>
+                </TouchableOpacity>
+              </View>
+            )}
+          /> */}
+        </View>
+        <TouchableOpacity
+          activeOpacity={0.7}
+          onPress={() => {
+            callback(username, token);
+          }}
+          style={styles.refreshTouchableOpacityStyle}
+        >
+          <FontAwesomeIcon icon={faArrowsRotate} color="#D6D5A8" size={35} />
+        </TouchableOpacity>
+      </>
     );
   } else {
     // En théorie ça ne dervait jamais s'afficher
@@ -168,14 +202,14 @@ export default function ProjectList({ username, token, userRole }) {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
     backgroundColor: "#1B2430",
     alignItems: "center",
     justifyContent: "center",
-    paddingTop: 150,
+    paddingTop: 20,
   },
   text: {
     color: "#D6D5A8",
+    fontSize: 20,
   },
   link: {
     color: "#816797",
@@ -201,11 +235,7 @@ const styles = StyleSheet.create({
     width: 300,
     borderRadius: 10,
   },
-  text_item: {
-    marginLeft: 10,
-    width: 150,
-  },
-  touchableOpacityStyle: {
+  plusTouchableOpacityStyle: {
     position: "absolute",
     width: 50,
     height: 50,
@@ -214,10 +244,34 @@ const styles = StyleSheet.create({
     right: 30,
     bottom: 30,
   },
+  refreshTouchableOpacityStyle: {
+    position: "absolute",
+    width: 50,
+    height: 50,
+    alignItems: "center",
+    justifyContent: "center",
+    right: 30,
+    bottom: 90,
+  },
   floatingButtonStyle: {
     resizeMode: "contain",
     width: 50,
     height: 50,
     //backgroundColor:'black'
+  },
+  gridView: {
+    marginTop: 10,
+    flex: 1,
+  },
+  itemContainer: {
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 5,
+    height: 75,
+  },
+  text_item: {
+    marginLeft: 5,
+    width: 150,
+    marginBottom: 20,
   },
 });
